IDPL location change, and fixes for build/test for Linux x86_64/64bit of INTEGER fields
1. IDPL location change
2. INTB-4434: fixes for build/test for Linux x86_64/64bit of INTEGER/NUMERIC field values
3. Adding isc_portable_integer() API
4. Misc. changes to Test/SimpleTest.cpp

+++ IscDbc/Engine.h
@@ -49,7 +49,7 @@ typedef wchar_t			WCHAR;
 #endif
 
 typedef unsigned char	UCHAR;
-typedef unsigned long	ULONG;
+typedef unsigned int	ULONG;
 
 enum LockType {
 	None,
