+++ IscDbc/IscConnection.cpp
@@ -4,7 +4,7 @@
  *     Developer's Public License Version 1.0 (the "License"); 
  *     you may not use this file except in compliance with the 
  *     License. You may obtain a copy of the License at 
- *     http://www.ibphoenix.com/main.nfs?a=ibphoenix&page=ibp_idpl.
+ *     https://www.firebirdsql.org/en/initial-developer-s-public-license-version-1-0/
  *
  *     Software distributed under the License is distributed on 
  *     an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, either 
@@ -964,7 +964,7 @@ public:
 
 typedef MList<CSchemaIdentifier> ListSchemaIdentifier;
 
-bool IscConnection::removeSchemaFromSQL( char *strSql, int lenSql, char *strSqlOut, long &lenSqlOut )
+bool IscConnection::removeSchemaFromSQL( char *strSql, int lenSql, char *strSqlOut, int &lenSqlOut )
 {
 	ListSchemaIdentifier listSchemaIdentifierAll;
 	ListSchemaIdentifier listSchemaIdentifierTbl;
@@ -1220,9 +1220,9 @@ bool IscConnection::removeSchemaFromSQL( char *strSql, int lenSql, char *strSqlO
 	return success;
 }
 
-int IscConnection::getNativeSql (const char * inStatementText, long textLength1,
-								char * outStatementText, long bufferLength,
-								long * textLength2Ptr)
+int IscConnection::getNativeSql (const char * inStatementText, int textLength1,
+								char * outStatementText, int bufferLength,
+								int * textLength2Ptr)
 {
 	int statysModify = 0;
 	char * ptIn = (char*)inStatementText;
