+++ IscDbc/IscConnection.h
@@ -4,7 +4,7 @@
  *     Developer's Public License Version 1.0 (the "License"); 
  *     you may not use this file except in compliance with the 
  *     License. You may obtain a copy of the License at 
- *     http://www.ibphoenix.com/main.nfs?a=ibphoenix&page=ibp_idpl.
+ *     https://www.firebirdsql.org/en/initial-developer-s-public-license-version-1-0/
  *
  *     Software distributed under the License is distributed on 
  *     an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, either 
@@ -114,10 +114,10 @@ public:
 	JString getInfoString (char *buffer, int item, const char *defaultString);
 	int getInfoItem (char *buffer, int item, int defaultValue);
 	JString getIscStatusText (ISC_STATUS *statusVector);
-	bool removeSchemaFromSQL( char *strSql, int lenSql, char *strSqlOut, long &lenSqlOut );
-	virtual int getNativeSql (const char * inStatementText, long textLength1,
-								char * outStatementText, long bufferLength,
-								long * textLength2Ptr);
+	bool removeSchemaFromSQL( char *strSql, int lenSql, char *strSqlOut, int &lenSqlOut );
+	virtual int getNativeSql (const char * inStatementText, int textLength1,
+								char * outStatementText, int bufferLength,
+								int * textLength2Ptr);
 	isc_tr_handle startTransaction();
 	void deleteStatement (IscStatement *statement);
 	IscConnection();
