+++ OdbcStatement.cpp
@@ -4,7 +4,7 @@
  *     Developer's Public License Version 1.0 (the "License"); 
  *     you may not use this file except in compliance with the 
  *     License. You may obtain a copy of the License at 
- *     http://www.ibphoenix.com/main.nfs?a=ibphoenix&page=ibp_idpl.
+ *     https://www.firebirdsql.org/en/initial-developer-s-public-license-version-1-0/
  *
  *     Software distributed under the License is distributed on 
  *     an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, either 
@@ -394,8 +394,8 @@ SQLRETURN OdbcStatement::sqlPrepare(SQLCHAR * sql, int sqlLength)
 	{
 		if ( noscanSQL == SQL_NOSCAN_OFF )
 		{
-			long lenstrSQL = (long)strlen(string);
-			long lennewstrSQL = lenstrSQL + 4096;
+			int lenstrSQL = (int)strlen(string);
+			int lennewstrSQL = lenstrSQL + 4096;
 			
 			retNativeSQL = connection->connection->getNativeSql( string, lenstrSQL, tempNative. getBuffer( lennewstrSQL ), lennewstrSQL, &lenstrSQL );
